+++ site/src/libs/path.ts
@@ -13,11 +13,11 @@ export function getChassisAssetsFsPath() {
 }
 
 export function getChassisTokensFsPath() {
-  return path.join(process.cwd(), 'node_modules/@ozgurgunes/chassis-tokens/dist/tokens/web/chassis-docs')
+  return path.join(process.cwd(), 'node_modules/@chassis-ui/tokens/dist/tokens/web/chassis-docs')
 }
 
 export function getChassisCSSFsPath() {
-  return path.join(process.cwd(), 'node_modules/@ozgurgunes/chassis-css/dist')
+  return path.join(process.cwd(), 'node_modules/@chassis-ui/css/dist')
 }
 
 export function getChassisIconsFsPath() {